Proyecto con Angular completado

+++ aprendiendo-angular/src/app/components/article/article.component.ts
@@ -3,6 +3,7 @@ import { ArticleService } from '../../services/article.service';
 import { Article } from '../../models/article';
 import { Route, ActivatedRoute, Params, Router } from '@angular/router';
 import { global } from '../../services/global';
+import swal from 'sweetalert';
 
 @Component({
   selector: 'app-article',
@@ -42,15 +43,31 @@ export class ArticleComponent implements OnInit {
   }
 
   delete(id) {
-    this._articleService.delete(id).subscribe(
-      response => {
-        this._router.navigate(['/blog']);
-      },
-      error => {
-        console.log(error);
-        this._router.navigate(['/blog']);
-      }
-    );
+    swal({
+      title: "Estas seguro?",
+      text: "Una ves borrado el articulo no podras recuperarlo",
+      icon: "warning",
+      buttons: [true,true],
+      dangerMode: true
+    })
+      .then((willDelete) => {
+        if (willDelete) {
+          this._articleService.delete(id).subscribe(
+            response => {
+              swal("Articulo eliminado correctamente", {
+                icon: "success",
+              });
+              this._router.navigate(['/blog']);
+            },
+            error => {
+              console.log(error);
+              this._router.navigate(['/blog']);
+            }
+          );
+        } else {
+          swal("Se ha cancelado la eliminacion.");
+        }
+      });
   }
 
 }
